Move inputs from forms/inputs to inputs
