+++ src/app/app.module.ts
@@ -68,8 +68,8 @@ import {SkillProxy} from './services/proxy/skill-proxy.service';
 import {SliderComponent} from './components/slider/slider.component';
 import {SuccessMessageComponent} from './components/messages/success-message/success-message.component';
 import {SystemLanguagesService} from './services/system-languages.service';
-import {TextareaInputComponent} from './components/forms/inputs/textarea-input/textarea-input.component';
-import {TextInputComponent} from './components/forms/inputs/text-input/text-input.component';
+import {TextareaInputComponent} from './components/inputs/textarea-input/textarea-input.component';
+import {TextInputComponent} from './components/inputs/text-input/text-input.component';
 import {TranslationService} from './services/translation.service';
 import {TruncatePipe} from './utils/truncate';
 import {UserAddressInputComponent} from './components/users/inputs/user-address-input.component';
